backend/requirements.txt
Flask==3.1.0
pymongo==4.10.1
python-dotenv
requests
twilio
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
paddleocr
paddlepaddle
Pillow
easyocr
pytesseract
